Move schema.py into RimSort.utils
