+++ util/metadata.py
@@ -30,7 +30,7 @@ from RimSort.utils.constants import (
     RIMWORLD_DLC_METADATA,
 )
 from RimSort.utils.generic import directories
-from util.schema import validate_mods_config_format
+from RimSort.utils.schema import validate_mods_config_format
 from util.steam.steamcmd.wrapper import SteamcmdInterface
 from util.steam.steamfiles.wrapper import acf_to_dict, dict_to_acf
 from util.steam.webapi.wrapper import (
